requirements.txt
streamlit==1.28.1
pycryptodome==3.18.0
pillow==10.0.1
numpy==1.24.3
scipy==1.11.4
pandas==2.0.3
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2
pytest==7.4.3
scikit-image==0.21.0
